Adjust import for test file
It looks like we're installing the outer name correctly, but it's not seeing the inner name for importing.  That is different than I was seeing locally, but I'd also set my PYTHONPATH.  Try importing from the outer package name - if that works, great!  If not, I might need to do some __init__.py generation adjustment